+++ src/extension/backend/AlgorithmEnum.ts
@@ -1,6 +1,3 @@
 export enum AlgorithmEnum {
-    Boolean,
-    Vector,
-    LanguageModel,
     Fuzzy
 }
