--- src/extension/backend/algorithms/edit-distance/algo-enum.ts
@@ -1,4 +1,0 @@
-export enum AlgoEnum {
-    Damerau = "DL",
-    Levenhstein = "L"
-}
