--- src/extension/backend/results/RankedQueryResult.ts
@@ -1,6 +1,0 @@
-import QueryResult from "./QueryResult";
-
-export default interface RankedQueryResult extends QueryResult {
-  readonly relativeRank: number;
-  readonly score: number;
-}
